requirements-core.txt
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.0.0
tqdm>=4.62.0
